--- src/sysdep.rs
@@ -1,4 +1,0 @@
-pub mod arm;
-pub mod clang;
-pub mod compiler;
-pub mod x86;
