--- src/sysdep/arm.rs
@@ -1,1 +1,0 @@
-pub mod clang;
