--- src/sysdep/x86.rs
@@ -1,1 +1,0 @@
-pub mod clang;
